DiceRoller
This relatively easy project took a little time, but was very fun to code. For this dice roller, the user can change the number of sides to as many as he/she wants. This has many wide spread applications, but most commonly as a way to randomly decide between a group or list what is the action. I have used this dice roller on a few occasions in school.