Ad-Lib-Generator
A relatively simple project that I decided to do as a side project and a fuel to use coding to make a small game. I have also experimeted with various versions where user input will lead to different kids of stories or jokes. This will happen soon, will keep posted.