Decoder/Encoder with
This was by far - MY BIGGEST PROJECT. I took a very long time to fully understand how this functions, and I was very happy when it finally worked. For this, I used a keyword of something the user can use. so this keyword is then split into individual letters, and finally this shift in the key word is used to encode or decode a certain message. For example, with a keyword of friends, and a decoded message of "d ugrr qw rjlyu qpz osibra ondaifadxf oule aaaaaaa", if you use the decoder function of the project, you can find the real word behind the encoded one. Please try to use this with you friends as it will be a very fun tool.